requirements/base.txt
django==4.1
python-dotenv==0.21.0